GCDAProfiling: Fix growing mmapped gcda file
In the current code, if we open (and therefore mmap) an existing
file, we never grow it (either the file or the allocation/mapping).
We also don't check for running off the end, because that check is
done by the code that grows the buffer, which does nothing for an
mmap'd buffer.

If we need to grow the buffer larger than the mapped file, unmap
the file, malloc a buffer, and follow the malloc'd-buffer path
from then on, which will eventually overwrite the existing file.